[move/#76] RegisterResultType의 위치 변경
