+++ app/src/main/java/com/haphap/app/data/repository/impl/register/RegisterRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.haphap.app.data.mapper.register.toModel
 import com.haphap.app.data.model.register.RegisterDropDownItemModel
 import com.haphap.app.data.model.register.RegisterModel
 import com.haphap.app.data.model.register.RegisterProcessModel
-import com.haphap.app.data.model.register.RegisterResultType
+import com.haphap.app.presentation.register.type.RegisterResultType
 import com.haphap.app.data.model.register.RegistrationCheckModel
 import com.haphap.app.data.model.register.RegistrationModel
 import com.haphap.app.data.remote.datasource.api.register.RegisterDataSource
