+++ app/src/main/java/com/haphap/app/presentation/register/RegisterViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import com.haphap.app.data.model.register.RegisterDropDownItemModel
-import com.haphap.app.data.model.register.RegisterResultType
+import com.haphap.app.presentation.register.type.RegisterResultType
 import com.haphap.app.data.model.register.RegistrationCheckModel
 import com.haphap.app.data.repository.api.register.RegisterRepository
 import com.haphap.app.presentation.register.navigation.Register
